requirements.txt
fastapi
uvicorn
anthropic
aiofiles
python-dotenv
